requirements.txt
streamlit>=1.35.0
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.2.0
Pillow>=10.0.0
numpy>=1.26.0
pandas>=2.1.0